Move parser error to separate module
