+++ yash-syntax/src/parser/and_or.rs
@@ -16,11 +16,11 @@
 
 //! Syntax parser for and-or list
 
-use super::core::Error;
 use super::core::Parser;
 use super::core::Rec;
 use super::core::Result;
-use super::core::SyntaxError;
+use super::error::Error;
+use super::error::SyntaxError;
 use super::fill::MissingHereDoc;
 use super::lex::Operator::{AndAnd, BarBar};
 use super::lex::TokenId::Operator;
@@ -73,7 +73,7 @@ impl Parser<'_> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::core::ErrorCause;
+    use super::super::error::ErrorCause;
     use super::super::fill::Fill;
     use super::super::lex::Lexer;
     use super::*;
