+++ yash-syntax/src/parser/lex/arith.rs
@@ -17,9 +17,9 @@
 //! Part of the lexer that parses arithmetic expansions.
 
 use super::core::Lexer;
-use crate::parser::core::Error;
 use crate::parser::core::Result;
-use crate::parser::core::SyntaxError;
+use crate::parser::error::Error;
+use crate::parser::error::SyntaxError;
 use crate::source::Location;
 use crate::syntax::Text;
 use crate::syntax::TextUnit;
@@ -94,7 +94,7 @@ impl Lexer {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parser::core::ErrorCause;
+    use crate::parser::error::ErrorCause;
     use crate::source::Source;
     use crate::syntax::Backslashed;
     use crate::syntax::Literal;
