+++ yash-syntax/src/parser/lex/modifier.rs
@@ -19,9 +19,9 @@
 use super::core::Lexer;
 use super::core::WordContext;
 use super::core::WordLexer;
-use crate::parser::core::Error;
 use crate::parser::core::Result;
-use crate::parser::core::SyntaxError;
+use crate::parser::error::Error;
+use crate::parser::error::SyntaxError;
 use crate::syntax::Modifier;
 use crate::syntax::Switch;
 use crate::syntax::SwitchCondition;
@@ -150,7 +150,7 @@ impl WordLexer<'_> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parser::core::ErrorCause;
+    use crate::parser::error::ErrorCause;
     use crate::source::Source;
     use crate::syntax::Text;
     use crate::syntax::TextUnit;
